requirements.txt
# Essential packages for LightRAG project

# Core dependencies
lightrag-hku==1.3.0  # Main RAG framework
pymongo==4.13.0      # MongoDB connection
fastapi==0.115.12    # API framework
uvicorn==0.34.0      # ASGI server for FastAPI
tiktoken==0.9.0
pipmaster==0.5.4

# Environment and configuration
python-dotenv==1.0.0  # Environment variable management

# AI and ML dependencies
pydantic-ai==0.0.49   # Pydantic AI agent
pydantic==2.11.1      # Data validation
transformers==4.52.4 # Hugging Face Transformers for embedding models
google-genai==1.21.1

# Streamlit for UI
streamlit==1.44.1     # Web UI framework

# Utilities
tqdm==4.67.1          # Progress bars